+++ notification.go
@@ -3,13 +3,13 @@ package main
 import "log"
 
 type notificatorConfig struct {
-	Type   string
-	Target string
-	Params map[string]string
+	Type      string
+	Target    string
+	Params    map[string]string
 }
 
 type notificator interface {
-	Notify(amount uint, comment string) error
+	Notify(amount uint64, comment string) error
 	Target() string
 }
 
@@ -26,7 +26,7 @@ func setupNotificators(cfg Config) {
 	}
 }
 
-func broadcastNotification(amount uint, comment string) {
+func broadcastNotification(amount uint64, comment string) {
 	log.Printf("Received %d sats with comment: %s", amount, comment)
 	for _, n := range notificators {
 		err := n.Notify(amount, comment)
